switch to chef boxes for testing
ksplice won't work on the puppetlabs CentOS 7.0 box. I contacted ksplice about this and was told:

"""It looks like CentOS rebuilt the 3.10.0-123.el7.x86_64 after we added support for it but didn't modify the version.  We'll work on adding support for this old kernel, but in the meantime, updating to a later on-disk kernel and rebooting will allow you to start using Ksplice."""